Refactor libp2p sync and split types and tests
